+++ src/Player.cpp
@@ -1,6 +1,6 @@
 #include "CPlayer.hpp"
 
-CPlayer::CPlayer(string sName,string sPassword, string sID, int hp, size_t strength, size_t gold, CRoom* room, attacks newAttacks)
+CPlayer::CPlayer(string sName,string sPassword, string sID, int hp, size_t strength, int gold, CRoom* room, attacks newAttacks)
 {
     m_sName = sName;
     m_sPassword = sPassword;
